+++ notebooks/__init__.py
@@ -1,0 +1,4 @@
+"""
+Face Detection with Haar Cascade Notebooks
+Copyright (c) 2025, Ahmad Fadlilah (https://github.com/ahmadfadlilah)
+"""